Refactored the code.
