+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch/domain/model/AmJurisdictionItem.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.domain.model;
+package uk.gov.hmcts.reform.roleassignmentbatch.domain.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 @Builder(toBuilder = true)
 @AllArgsConstructor
-public class CcdRoleNameItem {
-    private String caseRole;
+public class AmJurisdictionItem {
+    private String caseTypeId;
     private long count;
 }
