+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch/domain/model/CcdJurisdiction.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.domain.model;
+package uk.gov.hmcts.reform.roleassignmentbatch.domain.model;
 
 import java.util.List;
 
@@ -9,5 +9,5 @@ import lombok.Data;
 @Builder(toBuilder = true)
 public class CcdJurisdiction {
     private long totalCcdJurisdictionsCount;
-    private List<uk.gov.hmcts.reform.domain.model.CcdJurisdictionItem> jurisdictions;
+    private List<CcdJurisdictionItem> jurisdictions;
 }
